Develop(helpers & server): Moved helper functions to helper files for clarity
